Update README.md
Since the UserVoice-page for Entity Framework was shuttered, it doesn't really make sense to continue linking users to voting for it. I opened a github issue on the EF Core repo instead - people could be sent there, or the sentence could be dropped completely. I opted for the first one.